+++ src/generate/renders/extraExt/CliExtReadme.ts
@@ -64,7 +64,7 @@ export class CliExtReadme extends TemplateBase {
                         const temp = CmdToMultiLines(example.CommandString);
                         exampleCommandList.push(...temp);
                     }
-                    const waitCommandString = this.model.GetExampleWait(example).join(' ');
+                    const waitCommandString = exampleHandler.GetExampleWait(example).join(' ');
                     if (!isNullOrUndefined(waitCommandString) && waitCommandString !== '') {
                         exampleCommandList.push('');
                         const temp = CmdToMultiLines(waitCommandString);
